+++ src/foam/dao/index/ValuePlan.java
@@ -25,6 +25,6 @@ public class ValuePlan implements FindPlan, SelectPlan {
     return null;
   }
 
-  public void select(Object state, Sink sink, int skip, int limit, Comparator order, Predicate predicate) {
+  public void select(Object state, Sink sink, long skip, long limit, Comparator order, Predicate predicate) {
   }
 }
